Refactor of Router configuration
